+++ core-bundle/src/Doctrine/Schema/DcaSchemaProvider.php
@@ -17,8 +17,10 @@ use Contao\Database\Installer;
 use Doctrine\Bundle\DoctrineBundle\Registry;
 use Doctrine\DBAL\Platforms\MySqlPlatform;
 use Doctrine\DBAL\Schema\Schema;
+use Doctrine\DBAL\Schema\SchemaConfig;
 use Doctrine\DBAL\Schema\Table;
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Tools\SchemaTool;
 
 class DcaSchemaProvider
@@ -73,6 +75,12 @@ class DcaSchemaProvider
                 foreach ($definitions['SCHEMA_FIELDS'] as $fieldName => $config) {
                     $options = $config;
                     unset($options['name'], $options['type']);
+
+                    // Use the binary collation if the "case_sensitive" option is set
+                    if ($this->isCaseSensitive($config)) {
+                        $options['platformOptions']['collation'] = $this->getBinaryCollation($table);
+                    }
+
                     $table->addColumn($config['name'], $config['type'], $options);
                 }
             }
@@ -148,7 +156,14 @@ class DcaSchemaProvider
      */
     private function createSchemaFromDca(): Schema
     {
-        $schema = new Schema();
+        $config = new SchemaConfig();
+        $params = $this->doctrine->getConnection()->getParams();
+
+        if (isset($params['defaultTableOptions'])) {
+            $config->setDefaultTableOptions($params['defaultTableOptions']);
+        }
+
+        $schema = new Schema([], [], $config);
 
         $this->appendToSchema($schema);
 
@@ -176,7 +191,8 @@ class DcaSchemaProvider
 
         $this->setLengthAndPrecisionByType($type, $dbType, $length, $scale, $precision, $fixed);
 
-        $type = $this->doctrine->getConnection()->getDatabasePlatform()->getDoctrineTypeMapping($type);
+        $connection = $this->doctrine->getConnection();
+        $type = $connection->getDatabasePlatform()->getDoctrineTypeMapping($type);
 
         if (0 === $length) {
             $length = null;
@@ -188,6 +204,11 @@ class DcaSchemaProvider
 
         if (preg_match('/collate ([^ ]+)/i', $def, $match)) {
             $collation = $match[1];
+        }
+
+        // Use the binary collation if the BINARY flag is set (see #1286)
+        if (0 === strncasecmp($def, 'binary ', 7)) {
+            $collation = $this->getBinaryCollation($table);
         }
 
         $options = [
@@ -387,8 +408,8 @@ class DcaSchemaProvider
     private function getMaximumIndexLength(Table $table, string $column): int
     {
         $indexLength = $this->getDefaultIndexLength($table);
+        $collation = $table->getOption('collate');
         $connection = $this->doctrine->getConnection();
-        $collation = $connection->getParams()['defaultTableOptions']['collate'];
 
         // Read the table collation if the table exists
         if ($connection->getSchemaManager()->tablesExist([$table->getName()])) {
@@ -437,5 +458,37 @@ class DcaSchemaProvider
         }
 
         return 767;
+    }
+
+    /**
+     * Checks if a field has the case-sensitive flag.
+     *
+     * @param array $config
+     *
+     * @return bool
+     */
+    private function isCaseSensitive(array $config): bool
+    {
+        if (!isset($config['customSchemaOptions']['case_sensitive'])) {
+            return false;
+        }
+
+        return true === $config['customSchemaOptions']['case_sensitive'];
+    }
+
+    /**
+     * Returns the binary collation depending on the charset.
+     *
+     * @param Table $table
+     *
+     * @return string|null
+     */
+    private function getBinaryCollation(Table $table): ?string
+    {
+        if (!$table->hasOption('charset')) {
+            return null;
+        }
+
+        return $table->getOption('charset').'_bin';
     }
 }