+++ core-bundle/tests/Doctrine/DoctrineTestCase.php
@@ -76,6 +76,7 @@ abstract class DoctrineTestCase extends TestCase
             ->willReturn(
                 [
                     'defaultTableOptions' => [
+                        'charset' => 'utf8mb4',
                         'collate' => 'utf8mb4_unicode_ci',
                     ],
                 ]